move RNN class to `baselines/ppo2/layers.py' and revert `baselines/common/models.py` to 858afa8.
