+++ baselines/common/models.py
@@ -1,32 +1,17 @@
 import numpy as np
 import tensorflow as tf
-import tensorflow.contrib.layers as layers
-
 from baselines.a2c import utils
 from baselines.a2c.utils import conv, fc, conv_to_fc, batch_to_seq, seq_to_batch
 from baselines.common.mpi_running_mean_std import RunningMeanStd
+import tensorflow.contrib.layers as layers
 
 mapping = {}
 
-
-def register(name, is_rnn=False):
+def register(name):
     def _thunk(func):
-        if is_rnn:
-            func = RNN(func)
         mapping[name] = func
         return func
-
     return _thunk
-
-
-class RNN(object):
-    def __init__(self, func, memory_size=None):
-        self._func = func
-        self.memory_size = memory_size
-
-    def __call__(self, *args, **kwargs):
-        return self._func(*args, **kwargs)
-
 
 def nature_cnn(unscaled_images, **conv_kwargs):
     """
@@ -61,7 +46,6 @@ def mlp(num_layers=2, num_hidden=64, activation=tf.tanh, layer_norm=False):
 
     function that builds fully connected network with a given input tensor / placeholder
     """
-
     def network_fn(X):
         h = tf.layers.flatten(X)
         for i in range(num_layers):
@@ -79,7 +63,6 @@ def mlp(num_layers=2, num_hidden=64, activation=tf.tanh, layer_norm=False):
 def cnn(**conv_kwargs):
     def network_fn(X):
         return nature_cnn(X, **conv_kwargs)
-
     return network_fn
 
 
@@ -94,11 +77,10 @@ def cnn_small(**conv_kwargs):
         h = conv_to_fc(h)
         h = activ(fc(h, 'fc1', nh=128, init_scale=np.sqrt(2)))
         return h
-
     return network_fn
 
 
-@register("lstm", is_rnn=True)
+@register("lstm")
 def lstm(nlstm=128, layer_norm=False):
     """
     Builds LSTM (Long-Short Term Memory) network to be used in a policy.
@@ -134,8 +116,8 @@ def lstm(nlstm=128, layer_norm=False):
 
         h = tf.layers.flatten(X)
 
-        M = tf.placeholder(tf.float32, [nbatch])  # mask (done t-1)
-        S = tf.placeholder(tf.float32, [nenv, 2 * nlstm])  # states
+        M = tf.placeholder(tf.float32, [nbatch]) #mask (done t-1)
+        S = tf.placeholder(tf.float32, [nenv, 2*nlstm]) #states
 
         xs = batch_to_seq(h, nenv, nsteps)
         ms = batch_to_seq(M, nenv, nsteps)
@@ -148,12 +130,12 @@ def lstm(nlstm=128, layer_norm=False):
         h = seq_to_batch(h5)
         initial_state = np.zeros(S.shape.as_list(), dtype=float)
 
-        return h, {'S': S, 'M': M, 'state': snew, 'initial_state': initial_state}
+        return h, {'S':S, 'M':M, 'state':snew, 'initial_state':initial_state}
 
     return network_fn
 
 
-@register("cnn_lstm", is_rnn=True)
+@register("cnn_lstm")
 def cnn_lstm(nlstm=128, layer_norm=False, **conv_kwargs):
     def network_fn(X, nenv=1):
         nbatch = X.shape[0]
@@ -161,8 +143,8 @@ def cnn_lstm(nlstm=128, layer_norm=False, **conv_kwargs):
 
         h = nature_cnn(X, **conv_kwargs)
 
-        M = tf.placeholder(tf.float32, [nbatch])  # mask (done t-1)
-        S = tf.placeholder(tf.float32, [nenv, 2 * nlstm])  # states
+        M = tf.placeholder(tf.float32, [nbatch]) #mask (done t-1)
+        S = tf.placeholder(tf.float32, [nenv, 2*nlstm]) #states
 
         xs = batch_to_seq(h, nenv, nsteps)
         ms = batch_to_seq(M, nenv, nsteps)
@@ -175,12 +157,12 @@ def cnn_lstm(nlstm=128, layer_norm=False, **conv_kwargs):
         h = seq_to_batch(h5)
         initial_state = np.zeros(S.shape.as_list(), dtype=float)
 
-        return h, {'S': S, 'M': M, 'state': snew, 'initial_state': initial_state}
+        return h, {'S':S, 'M':M, 'state':snew, 'initial_state':initial_state}
 
     return network_fn
 
 
-@register("cnn_lnlstm", is_rnn=True)
+@register("cnn_lnlstm")
 def cnn_lnlstm(nlstm=128, **conv_kwargs):
     return cnn_lstm(nlstm, layer_norm=True, **conv_kwargs)
 
@@ -213,9 +195,7 @@ def conv_only(convs=[(32, 8, 4), (64, 4, 2), (64, 3, 1)], **conv_kwargs):
                                            **conv_kwargs)
 
         return out
-
     return network_fn
-
 
 def _normalize_clip_observation(x, clip_range=[-5.0, 5.0]):
     rms = RunningMeanStd(shape=x.shape[1:])
