+++ baselines/ppo2/layers.py
@@ -2,12 +2,21 @@ import numpy as np
 import tensorflow as tf
 
 from baselines.a2c.utils import ortho_init, fc, lstm, lnlstm
-from baselines.common.models import register, nature_cnn, RNN
+from baselines.common.models import register, nature_cnn
 
 
-@register("ppo_lstm", is_rnn=True)
+class RNN(object):
+    def __init__(self, func, memory_size=None):
+        self._func = func
+        self.memory_size = memory_size
+
+    def __call__(self, *args, **kwargs):
+        return self._func(*args, **kwargs)
+
+
+@register("ppo_lstm")
 def ppo_lstm(num_units=128, layer_norm=False):
-    def _network_fn(input, mask, state):
+    def network_fn(input, mask, state):
         input = tf.layers.flatten(input)
         mask = tf.to_float(mask)
 
@@ -18,12 +27,12 @@ def ppo_lstm(num_units=128, layer_norm=False):
         h = h[0]
         return h, next_state
 
-    return RNN(_network_fn, memory_size=num_units * 2)
+    return RNN(network_fn, memory_size=num_units * 2)
 
 
-@register("ppo_cnn_lstm", is_rnn=True)
+@register("ppo_cnn_lstm")
 def ppo_cnn_lstm(num_units=128, layer_norm=False, **conv_kwargs):
-    def _network_fn(input, mask, state):
+    def network_fn(input, mask, state):
         mask = tf.to_float(mask)
         initializer = ortho_init(np.sqrt(2))
 
@@ -38,41 +47,32 @@ def ppo_cnn_lstm(num_units=128, layer_norm=False, **conv_kwargs):
         h = h[0]
         return h, next_state
 
-    return RNN(_network_fn, memory_size=num_units * 2)
+    return RNN(network_fn, memory_size=num_units * 2)
 
 
-@register("ppo_cnn_lnlstm", is_rnn=True)
+@register("ppo_cnn_lnlstm")
 def ppo_cnn_lnlstm(num_units=128, **conv_kwargs):
     return ppo_cnn_lstm(num_units, layer_norm=True, **conv_kwargs)
 
 
-@register("ppo_lstm_mlp", is_rnn=True)
+@register("ppo_lstm_mlp")
 def ppo_lstm_mlp(num_units=128, layer_norm=False):
-    def network_fn(input, mask):
-        memory_size = num_units * 2
-        nbatch = input.shape[0]
-        mask.get_shape().assert_is_compatible_with([nbatch])
-        state = tf.Variable(np.zeros([nbatch, memory_size]),
-                            name='lstm_state',
-                            trainable=False,
-                            dtype=tf.float32,
-                            collections=[tf.GraphKeys.LOCAL_VARIABLES])
+    def _network_fn(input, mask, state):
+        h = tf.layers.flatten(input)
+        mask = tf.to_float(mask)
 
-        def _network_fn(input, mask, state):
-            input = tf.layers.flatten(input)
-            mask = tf.to_float(mask)
+        if layer_norm:
+            h, next_state = lnlstm([h], [mask[:, None]], state, scope='lnlstm', nh=num_units)
+        else:
+            h, next_state = lstm([h], [mask[:, None]], state, scope='lstm', nh=num_units)
+        h = h[0]
 
-            h, next_state = lstm([input], [mask[:, None]], state, scope='lstm', nh=num_units)
-            h = h[0]
+        num_layers = 2
+        num_hidden = 64
+        activation = tf.nn.relu
+        for i in range(num_layers):
+            h = fc(h, 'mlp_fc{}'.format(i), nh=num_hidden, init_scale=np.sqrt(2))
+            h = activation(h)
+        return h, next_state
 
-            num_layers = 2
-            num_hidden = 64
-            activation = tf.nn.relu
-            for i in range(num_layers):
-                h = fc(h, 'mlp_fc{}'.format(i), nh=num_hidden, init_scale=np.sqrt(2))
-                h = activation(h)
-            return h, next_state
-
-        return state, RNN(_network_fn)
-
-    return RNN(network_fn)
+    return RNN(_network_fn, num_units * 2)
